+++ src/parser/base/builder/temp-grammar.ts
@@ -75,10 +75,11 @@ export class TempGrammarRule<
   NT: string;
   callback?: Callback<T, After, Ctx>;
   rejecter?: Rejecter<T, After, Ctx>;
+  rollback?: Callback<T, After, Ctx>;
 
   constructor(
     data: Partial<
-      Pick<TempGrammarRule<T, After, Ctx>, "callback" | "rejecter">
+      Pick<TempGrammarRule<T, After, Ctx>, "callback" | "rejecter" | "rollback">
     > &
       Pick<TempGrammarRule<T, After, Ctx>, "rule" | "NT">
   ) {