+++ src/parser/base/builder/utils/definition.ts
@@ -74,6 +74,7 @@ export function defToTempGRs<T, After, Ctx extends BaseParserContext<T, After>>(
           }),
           callback: ctx?.callback,
           rejecter: ctx?.rejecter,
+          rollback: ctx?.rollback,
         })
       );
     });